Finally move all parser function to gmp/parser.js
Move parse_int, parse_float and parse_yesno from gmp/utils.js to
gmp/parser.js.
