+++ ng/src/gmp/parser.js
@@ -27,22 +27,7 @@ import moment from 'moment';
 import {
   is_defined,
   is_empty,
-  parse_float,
-  parse_int,
-  parse_yesno,
-  NO_VALUE,
-  YES_VALUE,
 } from './utils.js';
-
-// export imported parse stuff from utils
-// this code will be moved at this place in future
-export {
-  parse_int,
-  parse_float,
-  parse_yesno,
-  YES_VALUE,
-  NO_VALUE,
-};
 
 export function parse_severity(value) {
   return is_empty(value) ? undefined : parse_float(value);
@@ -71,6 +56,34 @@ export function parse_text(text) {
     text_excerpt: '0',
   };
 }
+
+export function parse_int(value) {
+  const val = parseInt(value, 10);
+
+  if (isNaN(val)) {
+    return undefined;
+  }
+
+  return val;
+}
+
+export function parse_float(value) {
+  const val = parseFloat(value);
+
+  if (isNaN(val)) {
+    return undefined;
+  }
+
+  return val;
+}
+
+export const YES_VALUE = 1;
+export const NO_VALUE = 0;
+
+export function parse_yesno(value) {
+  return value === '1' || value === 1 ? YES_VALUE : NO_VALUE;
+}
+
 
 export function parse_csv(value) {
   if (is_empty(value)) {
