+++ ng/src/web/components/bar/severitybar.js
@@ -23,7 +23,8 @@
 
 import React from 'react';
 
-import {parse_float, is_defined} from 'gmp/utils.js';
+import {is_defined} from 'gmp/utils.js';
+import {parse_float} from 'gmp/parser.js';
 
 import PropTypes from '../../utils/proptypes.js';
 import {
