+++ ng/src/web/components/form/yesnoradio.js
@@ -25,7 +25,7 @@
 import React from 'react';
 
 import _ from 'gmp/locale.js';
-import {parse_yesno, YES_VALUE, NO_VALUE} from 'gmp/utils.js';
+import {parse_yesno, YES_VALUE, NO_VALUE} from 'gmp/parser.js';
 
 import PropTypes from '../../utils/proptypes.js';
