+++ ng/src/web/components/powerfilter/applyoverridesgroup.js
@@ -24,7 +24,8 @@
 import React from 'react';
 
 import _ from 'gmp/locale.js';
-import {is_defined, parse_int} from 'gmp/utils.js';
+import {is_defined} from 'gmp/utils.js';
+import {parse_int} from 'gmp/parser.js';
 
 import PropTypes from '../../utils/proptypes.js';
