+++ ng/src/web/pages/scanners/dialog.js
@@ -25,7 +25,8 @@
 import React from 'react';
 
 import _, {datetime} from 'gmp/locale.js';
-import {is_defined, filter, parse_int, select_save_id} from 'gmp/utils.js';
+import {is_defined, filter, select_save_id} from 'gmp/utils.js';
+import {parse_int} from 'gmp/parser.js';
 
 import PropTypes from '../../utils/proptypes.js';
 import {render_options} from '../../utils/render.js';
