+++ ng/src/web/pages/targets/dialog.js
@@ -25,7 +25,7 @@
 import React from 'react';
 
 import _ from 'gmp/locale.js';
-import {NO_VALUE} from 'gmp/utils.js';
+import {NO_VALUE} from 'gmp/parser.js';
 
 import PropTypes from '../../utils/proptypes.js';
 import {render_options} from '../../utils/render.js';
