+++ ng/src/web/pages/tasks/details.js
@@ -26,7 +26,8 @@ import React from 'react';
 import moment from 'moment';
 
 import _, {long_date} from 'gmp/locale.js';
-import {is_defined, YES_VALUE} from 'gmp/utils.js';
+import {is_defined} from 'gmp/utils.js';
+import {YES_VALUE} from 'gmp/parser.js';
 
 import {OPENVAS_SCAN_CONFIG_TYPE} from 'gmp/models/scanconfig.js';
 import {scanner_type_name} from 'gmp/models/scanner.js';
